+++ src/app.module.ts
@@ -6,6 +6,7 @@ import { AuthModule } from './auth/auth.module';
 import { CoursePartnerModule } from './course-partner/course-partner.module';
 import { CourseUserModule } from './course-user/course-user.module';
 import { CourseModule } from './course/course.module';
+import { CrispListenerModule } from './crisp-listener/crisp-listener.module';
 import { CrispModule } from './crisp/crisp.module';
 import { EventLoggerModule } from './event-logger/event-logger.module';
 import { FeatureModule } from './feature/feature.module';
@@ -45,6 +46,7 @@ import { WebhooksModule } from './webhooks/webhooks.module';
     EventLoggerModule,
     HealthModule,
     CrispModule,
+    CrispListenerModule,
   ],
 })
 export class AppModule {}
