+++ src/crisp/crisp.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import Crisp from 'crisp-api';
 import { sendMailchimpUserEvent } from 'src/api/mailchimp/mailchimp-api';
 import { MAILCHIMP_CUSTOM_EVENTS } from 'src/api/mailchimp/mailchimp-api.interfaces';
@@ -18,40 +18,9 @@ const CrispClient = new Crisp();
 const logger = new Logger('CrispLogger');
 
 @Injectable()
-export class CrispService implements OnModuleInit {
+export class CrispService {
   constructor(private eventLoggerService: EventLoggerService) {
     CrispClient.authenticateTier('plugin', crispPluginId, crispPluginKey);
-  }
-
-  onModuleInit() {
-    logger.log(`Crisp service initiated`);
-
-    try {
-      const handleCrispEvent = async (message, eventName) =>
-        await this.handleCrispEvent(message, eventName);
-
-      CrispClient.on('message:send', async function (message: CrispEventDto) {
-        handleCrispEvent(message, EVENT_NAME.CHAT_MESSAGE_SENT);
-      })
-        .then(function () {
-          logger.log('Crisp service listening to sent messages');
-        })
-        .catch(function (error) {
-          logger.error('Crisp service failed listening to sent messages:', error);
-        });
-
-      CrispClient.on('message:received', function (message: CrispEventDto) {
-        handleCrispEvent(message, EVENT_NAME.CHAT_MESSAGE_RECEIVED);
-      })
-        .then(function () {
-          logger.log('Crisp service listening to received messages');
-        })
-        .catch(function (error) {
-          logger.error('Crisp service failed listening to sent messages:', error);
-        });
-    } catch (error) {
-      logger.error('Crisp service failed to initiate:', error);
-    }
   }
 
   async handleCrispEvent(message: CrispEventDto, eventName: EVENT_NAME) {
@@ -71,6 +40,7 @@ export class CrispService implements OnModuleInit {
           sessionMetaData.email,
           MAILCHIMP_CUSTOM_EVENTS.CRISP_MESSAGE_RECEIVED,
         );
+        logger.log('Crisp service: CRISP_MESSAGE_RECEIVED event sent to mailchimp');
       }
     } catch (error) {
       throw new Error(`Failed to handle crisp event for ${eventName}: ${error}`);
